refactor: move StoryState into the story module
StoryState is story domain state, so it belongs in agents/story.py with the other story models rather than in report.py.
